requirements.txt
amqp==2.5.1
Babel==2.7.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
billiard==3.5.0.5
celery==4.0.0
certifi==2019.9.11
chardet==3.0.4
configparser==4.0.2
contextlib2==0.6.0.post1
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.1
defusedxml==0.6.0
Django==1.11.8
django-admin-interface==0.11.2
django-colorfield==0.1.15
django-filter==1.1.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-rest-swagger==2.2.0
djangorestframework==3.9.4
djangorestframework-csv==2.1.0
djangorestframework-simplejwt==4.1.4
djangorestframework-xml==1.4.0
drf-tracking==1.5.0
enum34==1.1.6
flower==0.9.3
futures==3.3.0
idna==2.8
importlib-metadata==0.23
ipython==5.8.0
ipython-genutils==0.2.0
itypes==1.1.0
Jinja2==2.10.3
kombu==4.6.5
MarkupSafe==1.1.1
more-itertools==5.0.0
numpy==1.16.5
openapi-codec==1.3.2
pandas==0.24.2
pathlib2==2.3.5
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==1.0.18
psycopg2==2.8.4
ptyprocess==0.6.0
Pygments==2.4.2
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
rest-framework-cache==0.1
scandir==1.10.0
simplegeneric==0.8.1
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.13.0
tornado==5.1.1
traitlets==4.3.3
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.25.6
vine==1.3.0
virtualenv==16.7.2
wcwidth==0.1.7
zipp==0.6.0
